Server: Allow bad GII val. count from UVNC Viewer
As of this writing, the UltraVNC Viewer code:

https://github.com/ultravnc/UltraVNC/blob/9fda383ba24e1cecf14193802600c3bcd7e6b1fc/vncviewer/vnctouch.cpp#L542

hard-codes rfbGIIValuatorEvent.count to 6 * rfbGIIValuatorEvent.first.
This is correct if the touch area is sent, which is usually the case,
but it is incorrect if the touch area is not sent.  The UltraVNC Server
ignores rfbGIIValuatorEvent.count, so this commit causes the TurboVNC
Server to print a one-time warning if that value is incorrect, rather
than disconnecting the viewer.