Viewer: Fix race condition in connection shutdown
If the viewer window is closed using its close gadget or a hotkey, then
CConn.close() will execute in the Swing Event Dispatch Thread (EDT), so
we need to ensure that CConnection.processMsg() cannot execute
concurrently in the RFB thread.  Otherwise, TightDecoder could try to
use the native TurboJPEG instance after the instance has been destroyed,
thus causing the Java process to crash with a segfault on Un*x and an
EXCEPTION_GUARD_PAGE error on Windows.