Build: Opt. disable TurboVNC Server init script
The TurboVNC Server init script is a legacy feature at this point, but
enough people use it that I don't want to abandon it.  However, I also
don't want to go to the trouble of migrating it to systemd until such is
absolutely necessary.  In order to prevent dependency failures on
rolling releases, such as openSUSE Tumbleweed, that are no longer
backward-compatible with init.d, this commit introduces a new CMake
variable (TVNC_BUILDINITSCRIPT) that can be used to disable building and
packaging the init script.

Addresses a concern raised in #319