JSch: Revert fbaadc2a
JSch doesn't actually support certificates.  Like OpenSSH, it can use
any key or certificate provided by an SSH agent, even if the key's
algorithm is not explicitly supported, as long as the algorithm is
listed in PubkeyAcceptedAlgorithms.  However, it is misleading to list
algorithms in the default value of PubkeyAcceptedAlgorithms if there is
no explicit support for them.